shared: move copy_missing_files_from_devtool into git_runner
The agent's guarded session now needs to restore devtool-tracked
tarball files (generated autotools output, secondary-tarball payloads)
after switching to the CVE branch, the same way the corrector does.
Move the helper and its _get_submodule_paths dependency from
cve_corrector.git_ops into shared.git_runner so both the corrector and
the agent can share one implementation instead of duplicating it.

- Relocate copy_missing_files_from_devtool + _get_submodule_paths to
  shared/git_runner.py, using run_capture in place of run_cmd_capture
- Re-export both from cve_corrector.git_ops (noqa F401) so existing
  corrector callers and tests keep working unchanged
- Point the copy-missing unit tests at shared.git_runner.run_capture
