+++ shared/git_runner.py
@@ -6,11 +6,14 @@ Provides two levels of abstraction:
 - run_capture(): low-level, returns CompletedProcess (for corrector)
 - run_git_stdout(): high-level git-only, returns stdout str (for agent)
 """
+import logging
 import subprocess
 from pathlib import Path
 from typing import Optional
 
 from shared import TEXT_ENCODING, TEXT_ERRORS, build_git_env
+
+logger = logging.getLogger(__name__)
 
 
 def is_git_cmd(cmd: list[str]) -> bool:
@@ -130,3 +133,87 @@ def run_git_display(args: list[str], cwd: Path) -> None:
         ['git', '--no-pager'] + args, cwd=cwd, env=build_git_env(),
         check=False
     )
+
+
+def _get_submodule_paths(workspace_path: Path) -> set[str]:
+    """Return registered submodule paths from .gitmodules, if any."""
+    gitmodules = workspace_path / '.gitmodules'
+    if not gitmodules.exists():
+        return set()
+    paths: set[str] = set()
+    for line in gitmodules.read_text().splitlines():
+        stripped = line.strip()
+        if stripped.startswith('path'):
+            # e.g. "path = modules/oniguruma"
+            _, _, value = stripped.partition('=')
+            if value.strip():
+                paths.add(value.strip())
+    return paths
+
+
+def copy_missing_files_from_devtool(workspace_path: Path) -> None:
+    """Copy files present in devtool but missing from the CVE branch.
+
+    Release tarballs contain generated autotools files (configure, Makefile.in,
+    m4/*.m4, etc.) and secondary-tarball payloads (e.g. libxml2's ``xmlconf/``
+    W3C conformance suite fetched via a second ``SRC_URI`` entry) that are
+    committed on the devtool branch but do not exist on the upstream-history
+    CVE branch. Switching to the CVE branch therefore strips them, and the
+    build (do_configure/do_compile) then fails on the missing files. This
+    copies them across from the devtool branch without tracking them, so the
+    build succeeds on the CVE branch too. It is a no-op when nothing is
+    missing (e.g. when already on the devtool branch).
+
+    Files under registered submodule paths are skipped — those are tracked
+    by git submodule and must not be overwritten with regular file copies.
+
+    Files under paths that HEAD tracks as symlinks are also skipped — release
+    tarballs dereference symlinks into real directories, and copying those
+    files would clobber the symlink and break subsequent cherry-picks.
+    """
+    devtool_files = run_capture(
+        ['git', 'ls-tree', '-r', '--name-only', 'devtool'], cwd=workspace_path)
+    if devtool_files.returncode != 0:
+        return
+    cve_files = run_capture(
+        ['git', 'ls-tree', '-r', '--name-only', 'HEAD'], cwd=workspace_path)
+    if cve_files.returncode != 0:
+        return
+
+    cve_set = set(cve_files.stdout.strip().splitlines())
+    submodule_paths = _get_submodule_paths(workspace_path)
+
+    # Collect paths that HEAD tracks as symlinks (git mode 120000). Release
+    # tarballs dereference symlinks into real directories, so the devtool
+    # branch lists the symlink target's contents (e.g.
+    # tutorial/swift/swift-dep/Sources/*.swift) as regular files. Copying
+    # those out of devtool would clobber the symlink with a real directory,
+    # leaving a staged deletion + untracked dir that blocks every subsequent
+    # cherry-pick. Skip anything living under a HEAD symlink: the symlink
+    # already resolves to an in-tree path, so nothing is actually missing.
+    symlink_prefixes: tuple[str, ...] = ()
+    cve_tree = run_capture(
+        ['git', 'ls-tree', '-r', 'HEAD'], cwd=workspace_path)
+    if cve_tree.returncode == 0:
+        symlinks = []
+        for line in cve_tree.stdout.strip().splitlines():
+            # Format: "<mode> <type> <hash>\t<path>"
+            meta, _, path = line.partition('\t')
+            if path and meta.split(' ', 1)[0] == '120000':
+                symlinks.append(path + '/')
+        symlink_prefixes = tuple(symlinks)
+
+    missing = [f for f in devtool_files.stdout.strip().splitlines()
+               if f not in cve_set
+               and not any(f == sm or f.startswith(sm + '/')
+                           for sm in submodule_paths)
+               and not (symlink_prefixes and f.startswith(symlink_prefixes))]
+    if not missing:
+        return
+
+    logger.info("Copying %s missing file(s) from devtool branch", len(missing))
+    for filepath in missing:
+        run_capture(['git', 'checkout', 'devtool', '--', filepath],
+                    cwd=workspace_path)
+    # Unstage so they remain as untracked working-tree files
+    run_capture(['git', 'reset', 'HEAD'] + missing, cwd=workspace_path)
